upload to backend db + GCS

+++ app/src/main/java/com/example/applayout/Data/Model/Model.kt
@@ -5,7 +5,7 @@ data class Model(
     var model_task: String,
     var last_trained: String,
     var description: String,
-    var weight_size: Int,
+    var weight_size: Long,
     var is_uploaded: Boolean,
     var usage: Int,
     var likes: Int,
